auth-bot/requirements.txt
asyncpg==0.29.0
python-telegram-bot==20.7 